# backend/docs/swagger.yaml
basePath: /
definitions:
...
  server.CreatePatientRequest:
    properties:
      address:
        description: Use pointer for optional field
        type: string
      age:
        type: integer
      birthdate:
        description: Expect YYYY-MM-DD string
        type: string
      email:
        type: string
      firstname:
        type: string
      gender:
        type: string
      lastname:
        type: string
      phonenumber:
        type: string
      register:
        type: string
    type: object
...
  server.HTTPError:
    properties:
      error:
        description: Error message
        example: Resource not found
        type: string
    type: object
...
  server.PatientDetailsResponse:
    properties:
      distinct_diseases_list:
        description: Parsed list from patient_disease
        items:
          type: string
        type: array
      email:
        type: string
      firstname:
        type: string
      general_symptoms_list:
        description: Parsed list from patient_symptoms
        items:
          type: string
        type: array
      lastname:
        type: string
      patient_id:
        type: integer
    type: object
...
  server.PatientResponse:
    properties:
      address:
        description: string or null
        type: string
      age:
        type: integer
      birthdate:
        description: YYYY-MM-DD or null
        type: string
      email:
        type: string
      firstname:
        type: string
      gender:
        type: string
      lastname:
        type: string
      patient_id:
        type: integer
      phonenumber:
        type: string
      register:
        type: string
    type: object
...
  server.UpdatePatientRequest:
    properties:
      address:
        type: string
      age:
        type: integer
      birthdate:
        description: Expect YYYY-MM-DD string
        type: string
      email:
        type: string
      firstname:
        type: string
      gender:
        type: string
      lastname:
        type: string
      phonenumber:
        type: string
      register:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API Service for managing patients, symptoms, and diseases.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Patient API Service
  version: "1.0"
paths:
...
  /patients:
    get:
      consumes:
      - application/json
      description: Get a paginated list of patients
      parameters:
      - default: 10
        description: Pagination limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Pagination offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved list of patients
          schema:
            items:
              $ref: '#/definitions/server.PatientResponse'
            type: array
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: List patients
      tags:
      - Patients
    post:
      consumes:
      - application/json
      description: Add a new patient record to the database
      parameters:
      - description: Patient data to create
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/server.CreatePatientRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Patient created successfully
          schema:
            $ref: '#/definitions/server.PatientResponse'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error (e.g., DB error)
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Create a new patient
      tags:
      - Patients
  /patients/{patientID}:
    delete:
      consumes:
      - application/json
      description: Delete a patient record by ID. Associated records in junction tables
        (patient_symptoms, patient_disease) should be removed via ON DELETE CASCADE.
      parameters:
      - description: Patient ID
        format: int32
        in: path
        name: patientID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content (Successful deletion)
          schema:
            type: string
        "400":
          description: Invalid Patient ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Delete patient
      tags:
      - Patients
    get:
      consumes:
      - application/json
      description: Retrieve details of a specific patient by their ID
      parameters:
      - description: Patient ID
        format: int32
        in: path
        name: patientID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved patient
          schema:
            $ref: '#/definitions/server.PatientResponse'
        "400":
          description: Invalid Patient ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Patient not found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Get patient by ID
      tags:
      - Patients
    put:
      consumes:
      - application/json
      description: Update details for an existing patient
      parameters:
      - description: Patient ID
        format: int32
        in: path
        name: patientID
        required: true
        type: integer
      - description: Patient data to update
        in: body
        name: patient
        required: true
        schema:
          $ref: '#/definitions/server.UpdatePatientRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Patient updated successfully
          schema:
            $ref: '#/definitions/server.PatientResponse'
        "400":
          description: Invalid request payload or validation error
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Patient not found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Update patient details
      tags:
      - Patients
  /patients/{patientID}/details:
    get:
      consumes:
      - application/json
      description: Retrieve patient details along with aggregated lists of their general
        symptoms and distinct diseases recorded. Uses the GetPatientSummary query.
      parameters:
      - description: Patient ID
        format: int32
        in: path
        name: patientID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved patient summary details
          schema:
            $ref: '#/definitions/server.PatientDetailsResponse'
        "400":
          description: Invalid Patient ID format
          schema:
            $ref: '#/definitions/server.HTTPError'
        "404":
          description: Patient not found
          schema:
            $ref: '#/definitions/server.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/server.HTTPError'
      summary: Get patient summary details
      tags:
      - Patients
...
schemes:
- http
- https
swagger: "2.0"